Fixed checking tokens.
I had pushed a broken attempt at supporting both CRLF and LF delmiters
on the checktokens/groups values. There was an issue with how I was
splitting apart the checktoken values, so a check without an IP wasn't
being parsed correctly. There was also an incorrect check for instances
where a check didn't have an IP, but the token generally would.